Added Scaling for cube, which can be used with the F and G keys.

+++ SFMLOpenGL/Matrix3.cpp
@@ -205,7 +205,7 @@ Matrix3 Matrix3::Translate(double dx, double dy)
 	return answer;
 }
 
-Matrix3 Matrix3::Scale(int dx, int dy)
+Matrix3 Matrix3::Scale(double dx, double dy)
 {
 	Matrix3 answer;
 	answer.A11 = (double)dx / 100;
@@ -279,7 +279,7 @@ Matrix3 Matrix3::RotationZ(double _angle)
 	return answer;
 }
 
-Matrix3 Matrix3::Scale3D(int dx)
+Matrix3 Matrix3::Scale3D(double dx)
 {
 	Matrix3 answer;
 	answer.A11 = (double)dx / 100;
